// convex/_generated/api.ts
export const api = {
// ...
  projects: {
    listByOwner: {} as FunctionReference<"query">,
    get: {} as FunctionReference<"query">,
    create: {} as FunctionReference<"mutation">,
    update: {} as FunctionReference<"mutation">,
    remove: {} as FunctionReference<"mutation">,
    listSharedWith: {} as FunctionReference<"query">,
  },
// ...
  generations: {
    listByProject: {} as FunctionReference<"query">,
    get: {} as FunctionReference<"query">,
    create: {} as FunctionReference<"mutation">,
    updateStatus: {} as FunctionReference<"mutation">,
    saveResults: {} as FunctionReference<"mutation">,
    remove: {} as FunctionReference<"mutation">,
    getComments: {} as FunctionReference<"query">,
    addComment: {} as FunctionReference<"mutation">,
  },
// ...
} as const;